Create RiskOfRain v.0.0.1
This is a prototype of Risk of Rain that only has basic movement.
However, the movement is really choppy. 
Do whatever you can with it, but I think doing something like Risk of Rain might take a lot of more work than we might be able to do. (NetLogo's limitations)
So, maybe propose a new idea to Genkina, like Space Invaders or Color Switch or something...\
Let me know your stance on this.